Rough POC of retrieving camera thumbnails
This basically works for its intended purpose but is not currently complete.
In particular, it ignores API fields that it is not interested in and has
almost no error handling, silently failing if, for instance, network
connectivity is not available.

Sensitive or configurable data is passed in via the following environment
variables:

WYZE_CAM_LIST - comma separated list of the cameras whose events you want
                to subscribe to
WYZE_ACCESS_TOKEN - a refresh token generated from the Wyze site at
                   https://developer-api-console.wyze.com/#/apikey/view
WYZE_HOME - path on the file system where file thumbnails will be
            downloaded; this must end with a slash
SLACK_OAUTH_TOKEN - oauth token with write permissions for the channel
                    to which images will be posted; this should begin
                    with "xoxb-"
SLACK_CHANNEL - plaintext name or ID of the channel to which you want
                images posted